global movement + ruff
