+++ tests/unit_tests/utils/test_registration.py
@@ -31,7 +31,7 @@ class MockBittensorLogging:
 @pytest.fixture
 def mock_bittensor_logging(monkeypatch):
     mock_logger = MockBittensorLogging()
-    monkeypatch.setattr("bittensor.utils.registration.logging", mock_logger)
+    monkeypatch.setattr("bittensor.utils.registration.registration.logging", mock_logger)
     return mock_logger
 
 
@@ -48,7 +48,7 @@ def test_lazy_loaded_torch__torch_installed(monkeypatch, mock_bittensor_logging)
 
 
 def test_lazy_loaded_torch__no_torch(monkeypatch, mock_bittensor_logging):
-    monkeypatch.setattr("bittensor.utils.registration._get_real_torch", lambda: None)
+    monkeypatch.setattr("bittensor.utils.registration.registration._get_real_torch", lambda: None)
 
     torch = LazyLoadedTorch()
